requirements.txt
torch
pandas
numpy
scikit-learn
matplotlib
python-telegram-bot
joblib